Add URI (e.g. as represented by QR-codes) spectra parsing/writing.
Moved the URI spectrum parsing/writing code from InterSpec to SpecUtils.
Required some refactoring, additional code, and cleanup, as well as adding tests, etc.
In addition to making code to decode URIs, or create URIs, also made it possible to read in or write files whos contents is the URI stuff.

URI specification available in document "Draft Specification for Representing Gamma Radiation Spectra in QR codes, v1.1" at:
    https://sandialabs.github.io/InterSpec/tutorials/references/spectrum_in_a_qr_code_uur_latest.pdf

+++ src/D3SpectrumExport.cpp
@@ -847,7 +847,7 @@ D3SpectrumChartOptions::D3SpectrumChartOptions()
       ostr << "],";
     }//
     
-    ostr << std::setprecision(oldprecision);
+    ostr << std::setprecision(static_cast<int>(oldprecision));
     
     
     // foreground y-point values
